sparql_select: Insert query results into knowledge base to allow delayed unification.
This patch stores the results of a sparql result in a new predicate called "sparql_results".

See the example below:

rule() :-
	sparql_select(SparqlQuery, Id),
	sparql_results(Id, Var1, Var2, Var3...).

The Id term provided for sparql_select can be a variable, in which case it is assigned a random value.
The Var1, Var2, ... are in order corresponding to the ?var1 ?var2 ?var3 in the query.
Third parameter (replacement vars) and fourth parameter (sparql ep) can still be provided to sparql_select.
In the future, using the same Id twice might cause the former results to be removed from the KB, or
there might be some new sparql_delete(Id) call.